app/modules/pipelinem/pipcommands/pipc/run_test.go
package pipc

import (
	"strings"
	"testing"

	"github.com/goatcms/goatcore/app"
	"github.com/goatcms/goatcore/app/mockupapp"
)

func TestRunner(t *testing.T) {
	t.Parallel()
	var (
		err         error
		mapp        *mockupapp.App
		bootstraper app.Bootstrap
	)
	if mapp, bootstraper, err = newApp(mockupapp.MockupOptions{
		Input: strings.NewReader(` `),
		Args:  []string{`appname`, `pip:run`, `--name=name`, `--body="testCommand"`},
	}); err != nil {
		t.Error(err)
		return
	}
	// test
	if err = bootstraper.Run(); err != nil {
		t.Error(err)
		return
	}
	if err = mapp.AppScope().Wait(); err != nil {
		t.Error(err)
		return
	}
	buffer := mapp.OutputBuffer()
	if !strings.Contains(buffer.String(), "output") {
		t.Errorf("expected 'output' and take '%s'", buffer.String())
		return
	}
}
